+++ packages/utils/src/utils/dates/dateFormatters.ts
@@ -1,4 +1,12 @@
 const dayOfWeekNames = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'];
+const periodValues: Record<string, number> = {
+  year: 31536000,
+  month: 2592000,
+  week: 604800,
+  day: 86400,
+  hour: 3600,
+  min: 60,
+};
 
 export const formatTime = (timestampMs: number) =>
   new Date(timestampMs).toLocaleTimeString('en-US', {
@@ -32,4 +40,22 @@ export const printDayOfTheWeek = (day: number) => {
     case 6:
       return 'Sat';
   }
+};
+
+export const getTimeAgo = (timeValue: string | number) => {
+  const date = new Date(timeValue).getTime();
+  if (Number.isNaN(date)) {
+    throw new Error('Invalid date');
+  }
+  const diff = Math.floor((Date.now() - date) / 1000);
+  let interval;
+  for (const key in periodValues) {
+    interval = Math.floor(diff / periodValues[key]);
+    if (interval >= 1) {
+      const pluralValue = interval > 1 ? 's' : '';
+      return `${interval} ${key}${pluralValue} ago`;
+    }
+  }
+
+  return 'Just Now';
 };
