+++ packages/utils/src/utils/dates/index.ts
@@ -1,2 +1,1 @@
 export * as dateFormatters from './dateFormatters';
-export * as timeAgo from './timeAgo';
