+++ packages/widgets/src/components/NotificationCard/NotificationCard.stories.tsx
@@ -3,7 +3,7 @@ import { Theme } from '@emotion/react';
 import { Button, IMapIcons, MainColors } from '@ssa-ui-kit/core';
 import { mainTheme } from '@ssa-ui-kit/core';
 import { NotificationCard } from './NotificationCard';
-import { NotificationCardProps } from './types';
+import { INotificationCardProps } from './types';
 
 const childrenWrapper = (theme: Theme) => `
   grid-column: 2 / span 2;
@@ -92,7 +92,7 @@ export default {
 } as Meta<typeof NotificationCard>;
 
 export const Default: StoryObj<typeof NotificationCard> = (
-  args: NotificationCardProps,
+  args: INotificationCardProps,
 ) => {
   return (
     <div css={{ padding: '20px', background: '#F2F4F7', borderRadius: '20px' }}>
